Exception handllers

+++ src/main/java/com/devsuperior/dscommerce/services/ProductService.java
@@ -2,11 +2,16 @@ package com.devsuperior.dscommerce.services;
 
 import com.devsuperior.dscommerce.DTO.ProductDTO;
 import com.devsuperior.dscommerce.entities.Product;
+import com.devsuperior.dscommerce.exceptions.DatabaseException;
+import com.devsuperior.dscommerce.exceptions.ResourceNotFoundException;
 import com.devsuperior.dscommerce.repositories.ProductRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -31,7 +36,8 @@ public class ProductService {
      // Read (Buscar um produto por ID)
      @Transactional(readOnly = true)
      public ProductDTO findById(Long id) {
-          Product product = repository.findById(id).get();
+          Product product = repository.findById(id)
+                  .orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado: " + id));
           return new ProductDTO(product);
      }
 
@@ -46,16 +52,27 @@ public class ProductService {
      // Update (Atualizar um produto existente)
      @Transactional
      public ProductDTO update(Long id, ProductDTO dto) {
-          Product product = repository.getReferenceById(id);
-          copyDtoToEntity(dto, product);
-          product = repository.save(product);
-          return new ProductDTO(product);
+          try {
+               Product product = repository.getReferenceById(id);
+               copyDtoToEntity(dto, product);
+               product = repository.save(product);
+               return new ProductDTO(product);
+          } catch (EntityNotFoundException e) {
+               throw new ResourceNotFoundException("ID não encontrado: " + id);
+          }
      }
 
      // Delete (Remover um produto por ID)
-     @Transactional
+     @Transactional(propagation = Propagation.SUPPORTS)
      public void delete(Long id) {
-          repository.deleteById(id);
+          if (!repository.existsById(id)) {
+               throw new ResourceNotFoundException("ID não encontrado: " + id);
+          }
+          try {
+               repository.deleteById(id);
+          } catch (DataIntegrityViolationException e) {
+               throw new DatabaseException("Violação de integridade");
+          }
      }
 
      // Método utilitário para copiar os dados do DTO para a entidade
